+++ apps/deploy-web/src/instrumentation.ts
@@ -1,4 +1,7 @@
+import { LoggerService } from "@akashnetwork/logging";
 import * as Sentry from "@sentry/nextjs";
+
+const logger = new LoggerService({ name: `instrumentation-${process.env.NEXT_RUNTIME}` });
 
 export async function register() {
   // Note: if you want to override the automatic release value, do not set a
@@ -13,12 +16,11 @@ export async function register() {
 
   if (process.env.NEXT_RUNTIME === "nodejs") {
     try {
-      await import("@akashnetwork/env-loader");
-      const { serverEnvSchema } = await import("./config/env-config.schema");
+      const [, { serverEnvSchema }] = await Promise.all([import("@akashnetwork/env-loader"), import("./config/env-config.schema")]);
 
       serverEnvSchema.parse(process.env);
     } catch (error) {
-      console.error("Failed to validate server environment variables", error);
+      logger.error({ message: "Failed to validate server environment variables", error });
       process.exit(1);
     }
   }