+++ apps/deploy-web/src/pages/alerts/notification-channels/[id]/index.tsx
@@ -1,22 +1,12 @@
 import type { components } from "@akashnetwork/react-query-sdk/notifications";
-import { getSession } from "@auth0/nextjs-auth0";
-import type { GetServerSideProps } from "next";
 import type { GetServerSidePropsResult } from "next/types";
 import { z } from "zod";
 
 import { EditNotificationChannelPage } from "@src/components/alerts/EditNotificationChannelPage";
-import type { ServerServicesContext } from "@src/lib/nextjs/getServerSidePropsWithServices";
-import { getValidatedServerSideProps, type ValidatedServerSideContext } from "@src/lib/nextjs/getValidatedServerSideProps";
-import { featureFlagService } from "@src/services/feature-flag";
-import { notificationsApi } from "@src/services/server-side-notifications-api/server-side-notifications-api.service";
+import { defineServerSideProps } from "@src/lib/nextjs/defineServerSideProps/defineServerSideProps";
+import { isFeatureEnabled, isRegisteredUser } from "@src/lib/nextjs/pageGuards/pageGuards";
 
 export default EditNotificationChannelPage;
-
-const contextSchema = z.object({
-  params: z.object({
-    id: z.string().uuid()
-  })
-});
 
 type Props = {
   notificationChannel: components["schemas"]["NotificationChannelOutput"]["data"];
@@ -26,22 +16,17 @@ const NOT_FOUND: GetServerSidePropsResult<Props> = {
   notFound: true
 };
 
-export const getServerSideProps: GetServerSideProps<Props> = getValidatedServerSideProps(
-  contextSchema,
-  async (context: ServerServicesContext & ValidatedServerSideContext<typeof contextSchema>) => {
-    const session = await getSession(context.req, context.res);
-
-    if (!session?.user) {
-      return NOT_FOUND;
-    }
-
-    const isEnabled = await featureFlagService.isEnabledForCtx("alerts", context, { userId: session?.user?.id });
-
-    if (!isEnabled) {
-      return NOT_FOUND;
-    }
-
-    const notificationChannel = await notificationsApi.v1.getNotificationChannel({
+export const getServerSideProps = defineServerSideProps({
+  route: "/alerts/notification-channels/[id]",
+  schema: z.object({
+    params: z.object({
+      id: z.string().uuid()
+    })
+  }),
+  if: async ctx => (await isRegisteredUser(ctx)) && (await isFeatureEnabled("alerts", ctx)),
+  handler: async (context): Promise<GetServerSidePropsResult<Props>> => {
+    const session = (await context.services.getSession(context.req, context.res))!;
+    const notificationChannel = await context.services.notificationsApi.v1.getNotificationChannel({
       parameters: {
         path: {
           id: context.params.id
@@ -62,4 +47,4 @@ export const getServerSideProps: GetServerSideProps<Props> = getValidatedServerS
       }
     };
   }
-);
+});